+++ Infrastructure/ValueTypes/OpenAiRequest.cs
@@ -5,7 +5,7 @@ namespace Infrastructure.ValueTypes;
 /// <remarks>
 /// <see href="https://platform.openai.com/docs/api-reference/chat/create">OpenAI API Reference</see>
 /// </remarks>
-public record OpenAiRequest(string Model, OpenAiMessage[] Messages, double Temperature)
+public record OpenAiRequest(string Model, OpenAiMessage[] Messages, decimal Temperature)
 {
     public static OpenAiRequest Create(string modelId, TextPrompt prompt, PromptSettings promptSettings)
     {
